Extend and refine table/column functionality

+++ src/main/kotlin/org/incava/mmonkeys/rand/RandIntCalculated.kt
@@ -4,7 +4,7 @@ import kotlin.random.Random
 
 class RandIntCalculated(size: Int, numSlots: Int) : RandCalculated(size, numSlots) {
     private val random = Random.Default
-    private val intSlots: Map<Int, Int> = slots.mapValues { it.value.toInt() }
+    private val intSlots: Map<Int, Int> = slots.mapValues { it.value }
     private val intArray: List<Int> = intSlots.toSortedMap().map { it.value }
 
     fun nextMapInt() : Int {
